chore: refactor completions and definitions
