+++ server/src/utils/completions/index.ts
@@ -1,0 +1,6 @@
+export * from "./types/attributeModifier";
+export * from "./types/attributeName";
+export * from "./types/closeTag";
+export * from "./types/openTag";
+export * from "./types/openTagName";
+export * from "./types/styleContent";
