+++ server/src/utils/completions/types/attributeName.ts
@@ -1,0 +1,99 @@
+import {
+  CompletionParams,
+  CompletionList,
+  CompletionItem,
+  CompletionItemKind,
+  MarkupKind,
+  MarkupContent,
+  InsertTextFormat,
+  TextDocument,
+  TextEdit
+} from "vscode-languageserver";
+import { ParserEvents } from "../../htmljs-parser";
+import { TagLibLookup } from "../../compiler";
+import { rangeFromEvent } from "../../utils";
+
+export function attributeName(
+  taglib: TagLibLookup,
+  document: TextDocument,
+  params: CompletionParams,
+  event: ParserEvents.AttributeName
+) {
+  const completions: CompletionItem[] = [];
+  const attrNameRange = rangeFromEvent(document, event);
+  const tagDef =
+    !event.tag.tagNameExpression && taglib.getTag(event.tag.tagName);
+  const nestedTagAttrs: { [x: string]: boolean } = {};
+
+  if (tagDef && tagDef.nestedTags) {
+    for (const key in tagDef.nestedTags) {
+      const nestedTagDef = tagDef.nestedTags[key];
+      nestedTagAttrs[nestedTagDef.targetProperty] = true;
+    }
+  }
+
+  taglib.forEachAttribute((tagDef && tagDef.name) || "*", (attr, parent) => {
+    if (
+      attr.deprecated ||
+      nestedTagAttrs[attr.name] ||
+      attr.name === "*" ||
+      (attr.name[0] === "_" &&
+        /\/node_modules\//.test(attr.filePath || parent.filePath))
+    ) {
+      return;
+    }
+
+    const type = attr.type || (attr.html ? "string" : null);
+    let label = attr.name;
+    let snippet = attr.name;
+    let documentation: MarkupContent | undefined;
+
+    if (attr.enum) {
+      snippet += `="\${1|${attr.enum.join()}|}"$0`;
+    } else {
+      switch (type) {
+        case "string":
+          snippet += '="$1"$0';
+          break;
+        case "function":
+          snippet += "=($1)$0";
+          break;
+        case "statement":
+        case "boolean":
+        case "flag":
+          break;
+        default:
+          snippet += "=";
+          break;
+      }
+    }
+
+    const autocomplete =
+      attr.autocomplete && Array.isArray(attr.autocomplete)
+        ? attr.autocomplete[0]
+        : attr.autocomplete;
+
+    if (autocomplete) {
+      label = autocomplete.displayText || label;
+      snippet = autocomplete.snippet || snippet;
+
+      if (autocomplete.descriptionMoreURL) {
+        documentation = {
+          kind: MarkupKind.Markdown,
+          value: `[More Info](${autocomplete.descriptionMoreURL})`
+        };
+      }
+    }
+
+    completions.push({
+      label,
+      documentation,
+      detail: attr.description,
+      kind: CompletionItemKind.Property,
+      insertTextFormat: InsertTextFormat.Snippet,
+      textEdit: TextEdit.replace(attrNameRange, snippet)
+    });
+  });
+
+  return CompletionList.create(completions);
+}
