+++ server/src/utils/definitions/index.ts
@@ -1,0 +1,2 @@
+export * from "./types/attributeName";
+export * from "./types/openTagName";
